Error: samples_selected doesn't update for D optimality
Python function arguments are passed by assignment. (pass the vlaue of reference) When the argument is
mutable, like in this case sampels_selected is a list, if operated on
the list, like lst.append(), it will change the outloop values together
(bc it is the reference appending values)

When it is reasign to other value, lst=list2. What happend in the new
lst will not affect values in outer loop.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   experiment/optimal_design.py